+++ MoodleAssistant/Logic/Processing/Aiken/NamespaceDoc.cs
@@ -1,0 +1,6 @@
+﻿namespace MoodleAssistant.Logic.Processing.Aiken;
+
+/// <summary>
+/// Contains all the classes to handle Aiken template questions.
+/// </summary>
+internal static class NamespaceDoc{ }