+++ CHANGELOG.md
@@ -6,9 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+* **Logic**: added Aiken question type support
+
 ### Changed
 
-* **Logic**: minor refactoring
+* **Logic**: namespace refactoring, aiken merger refactor
 
 ## [0.9.0] - 2025-02-21